Move lament into module, remove progress bar
In preparation for using lament as part of my portfolio website, I've
moved it into a folder, along with its static files and templates, as
well as its submodules. Later, I'll use a wrapper to serve lament as part
of the portfolio website (or possibly use Flask blueprints to register it
on the main app).
I've also removed all files and code related to the progress
bar, as I'm no longer using it.